Refactor code by rearranging
